feature: set ResponseEmitter encoding type
Add the ability to explicitly set the response emitter encoding type.

This is used to set the encoding type of the response emitter to something other than the default. For example, when setting the encoding type to `Gzip` when compressed output is being emitted.

Needed for: https://github.com/ipfs/kubo/issues/2376